+++ src/adrpy/injection/__init__.py
@@ -1,9 +1,6 @@
-from lidipy import Lidi
-
+from adrpy.injection.container import lidi
 from adrpy.injection.modules import bind_modules
 from adrpy.injection.settings import bind_settings
-
-lidi = Lidi()
 
 
 def setup_injection() -> None: