+++ src/adrpy/injection/settings.py
@@ -1,7 +1,7 @@
 from lidipy import Lidi
 
+from adrpy.shared_kernel.settings import Settings
+
 
 def bind_settings(lidi: Lidi) -> None:
-    from adrpy.shared_kernel.settings import Settings
-
     lidi.bind(Settings, Settings(), singleton=True)