+++ tests/conftest.py
@@ -4,6 +4,7 @@ from typing import Generator, Iterator
 import pytest
 from lidipy import Lidi
 
+from adrpy.injection import lidi as injection_lidi
 from adrpy.injection import setup_injection
 from tests.fixtures.repository import TEST_DIRECTORY
 
@@ -12,10 +13,8 @@ pytest_plugins = ["fixtures.repository"]
 
 @pytest.fixture
 def lidi() -> Generator[Lidi, None, None]:
-    from adrpy.injection import lidi
-
     setup_injection()
-    yield lidi
+    yield injection_lidi
 
 
 @pytest.fixture(autouse=True)